style(marimekko): Split out type definitions, rearrange code
